Updated "How to configure sandbox environment"
I renamed the client variable to defaultClient in the "How to configure sandbox environment" code sample to keep consistency with the name given to the SquareConnect.ApiClient.instance in the first "Getting Started" code sample. Because in my case, I copied and pasted the first example and then just copied the line that sets the basePath, and because the naming is different, my code was failing. Hopefully, this fix in the documentation will allow distracted copy-pasters to move along seamlessly.